Upgrade to scala-2.12.15 (WIP)

+++ org.scala-ide.sdt.core/src/org/scalaide/ui/internal/editor/hover/BrowserControlAdditions.scala
@@ -29,9 +29,8 @@ trait BrowserControlAdditions extends BrowserInformationControl with HasLogger {
     gc.setFont(JFaceResources.getFontRegistry().get(IScalaHover.HoverFontId))
     val averageCharWidth = gc.getFontMetrics().getAverageCharacterWidth()
     gc.dispose()
-
     val p = super.computeSizeHint()
-    p.x += averageCharWidth * 10
+    p.x += (averageCharWidth * 10).toInt
     p
   }
 }
